Minor Update
Added in some of the artwork that Colin has done. Some of it will
require revisions for sizing but it definitely looks nice. I create a
small placeholder and template for the notification system. I already
have an idea of how it's going to work, so it should not take long when
I come back to it later.